feat(entity): add Constructor getter and setter

+++ Hospital-managment-system_Spring/src/main/java/com/shema/Hospital_managment_system_Spring/entity/MedicalInventory.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MedicalInventory {
- private Long itemId;
+ private int itemId;
  private String itemName;
  private int StockQuantity;
  private double unitPrice;
